third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/send-object-tostring-check-expected.txt

Test to verify XMLHttpRequest.send(...), with an object, calls toString exactly once.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testObj.callCounter is 1
PASS successfullyParsed is true

TEST COMPLETE
